Upload column：《MCU开发学习记录10》
feat: Upload column code

The PF2 level is controlled by KEY0, and the brake pin of TIM1 
is connected (valid for high level) to control the complementary output 
of ch1/ch1N of TIM1. At the same time, the complementary output of 
ch1/ch1N of TIM1 is configured with a dead time of 1ms.